Generate smote in pixels level and iterates through the best fscore

+++ smote.cpp
@@ -12,7 +12,7 @@
 void SMOTE::populate(Mat minority, Mat neighbors, Mat *synthetic, int *index, int amountSmote, int i, int nearestNeighbors){
 
     int attributes, nn, attr;
-    float gap, dif, attrOriginal, neighbor;
+    double gap, dif, attrOriginal, neighbor;
     vector<int> vectorRand;
     Size s = minority.size();
     attributes = s.width;
@@ -29,7 +29,7 @@ void SMOTE::populate(Mat minority, Mat neighbors, Mat *synthetic, int *index, in
                 neighbor = neighbors.at<float>(i, nn);
                 attrOriginal = minority.at<float>(i, attr);
 
-                /* The difference between the feature vector under and its 
+                /* The difference between the feature vector under and its
                 nearest neighbor*/
                 dif = minority.at<float>(neighbor, attr) - attrOriginal;
                 /* Multiply this difference with a number between 0 and 1 */
@@ -52,7 +52,7 @@ void SMOTE::computeNeighbors(Mat minority, int nearestNeighbors, Mat *neighbors)
     Mat classes(s.height, 1, CV_32FC1);
 
     for(i = 0; i < s.height; i++) {
-        classes.at<float>(i,0) = index; 
+        classes.at<float>(i,0) = index;
         index++;
     }
 
@@ -75,7 +75,7 @@ Mat SMOTE::smote(Mat minority, int amountSmote, int nearestNeighbors){
     vector<int> vectorRand;
     Mat newMinority;
     srand(time(0));
-    
+
     if(amountSmote == 0)
         return Mat();
